vertical-full-page-map: Add `alternativeVerticalsConfig` passthrough (#696)
Previously, the recommended other alternative verticals would show as a
suggestion, even if the query was an empty string. We have opted not to
show the vertical suggestions at all.

However, this is not toggle-able by the component in the SDK. In order
to remove that HTML, we must override the template to hide the HTML.

In order to pass a translated hbs file to the Alternative Verticals child of
the Vertical Full Page Map, we add an alternative verticals config
option to the Vertical Full Page Map component. This is better than just
a specific alternative verticals template config option.

We copy the alternativeverticals.hbs directly from the Answers Search
UI SDK except we add an exception for no-query suggested verticals. We
also escapeHTML all the translated files and copy the translations directly
from the Answers Search UI SDK .po files.

Note: An unused config option searchbarConfig was replaced with
this alternativeVerticalsConfig option.

J=SLAP-1203
TEST=manual

Test that when you get to a no results state on the new map, you will
not see alternative verticals show for any query.

Test that on the benchmark account, if you zoom into a search this
area search with no results (and no query) you do not get alternative
verticals. However, if you type e.g. wallet, you will get suggested
verticals to products (given you added the products vertical).