vertical-full-page-map: Move enablePinClustering option up (#701)
Move the option up because we want to uncomment this option easier. This
way, we do not need to remember to add a comma to the `pin` field and to
remove the comma in the enablePinClustering field.

J=None
TEST=manual

Test that you can import this page, uncomment the enablePinClustering,
and see pin clustering on the resulting page.