move "search this area" button below no results block
This commit makes the "search this area" button visible
during the no results state. Before this commit, the no
results block would cover up "search this area".

This was done by moving "search this area" directly
below Answers-resultsWrapper in the html.
In order for this to work, the css for "search this area"
was changed to use static positioning on mobile.

getBoundingClientRect() was used to calculate the new
margin-top for Answers-centerTop

J=SLAP-1225
TEST=manual

open up a side by side of a maps page before and after this commit
see that on mobile and desktop, the "search this area" button is in
the same position
see that I can swap back and forth between list and map view on mobile
see that in the no results state, the "search this area" button is directly
below the no results block, instead of being covered up by it