vertical-full-page-map: Fix panning map on initial results (#702)
Previously, on the initial search of the map, the results would not show
on the map if there was a populated search on load. This is because the
num concurrent search this area calls storage key wouldn't exist and
would give undefined instead of 0. We set a default for when the key
doesn't exist so we can check a number range.

This suggests we should re-name updateZoom to fitCoordinates as with
this bug, we were seeing both the coordinates not fit and the zoom not
changing.

J=None
TEST=manual

Test that you can see the results on a refresh of the site with a query
or a default initial search of ''.

Test that coordinates are not fit when you Search This Area or zoom
into a cluster.