dynamically resize the map on no results
This commit dynamically resizes the map when the no results
state appears/disappears. This is because the no results info
that pops up would otherwise cover up part of the map. This
would be a problem primarily if somebody had displayAllResults
set to true for the map, in which case a majority of the pins
would be blocked by alternative verticals suggestions.

J=SLAP-1237
TEST=manual

on no results, the map dynamically resizes
refreshing the page while on no results goes to the list
view, and the map is not visible
desktop looks correct for no results and regular results
"search this area" button is visible in no results