Fix logic for exiting the active state of a pin (#700)
Make it such that the active pin is deactivated if a user selects the search bar, or if a new search is ran

J=SLAP-1208
TEST=manual

Test that clicking on the search bar closes the detail card on mobile. Test that the detail card closes during "Search When Map Moves", and after pressing the "Search This Area" button. Test that the detail card closes when tabbing onto the search bar.