vertical-full-page-map: Fix Alternative Verts always showing (#708)
Because the `all` hbs helper uses JavaScript falsy values, we need to
change how we look at the `verticalSuggestions` object. The effect of
this is that alternative verticals messaging would show when the
verticalSuggestions array is empty.

Instead, we pass the `all` helper `verticalSuggestions.length` which hbs
defaults to undefined if `verticalSuggestions` is undefined so we can be
sure the array always has a length.

J=None
TEST=manual

Test on an experience with vertical suggestions and without that you see
them correctly.

I tested with the query `wallet` for our test account and `askjdhaksjd`
to see alternative verticals messaging show in the former and not the
latter.