Map list view fix for tall mobile screens (#703)
Fix a bug where the map would extend below the results on mobile.

This bug occurred on the list view page only within the mobile
breakpoint, and only on non-iframed experiences.

Fix the bug by adding a new YxtPage-wrapper--mobileListView CSS
class which lets us apply a fix to the Answers-footer. We fix the footer
by removing the margin using padding instead so that the white
background of the footer covers the map.

This PR updates the map orchestrator to improve the handling of the
mobileListView and the mobileMapView CSS classes. These changes
ensure that classes associated with mobile views only exist when the
page is within the mobile breakpoint. This has the side effect of fixing
layout issues when someone drags the screen between mobile and
desktop sizes.

J=SLAP-1230
TEST=manual

Test very tall screen sizes within the mobile breakpoint on list view
and confirm that the bug is fixed. Test on iframe and non-iframe
experiences to make sure things work as expected. Observe the
listShown CSS class being properly set on the page. Ensure that
a YxtFooter can be put on the page and have it look as expected.