vertical-full-page-map: Support no results display all results (#710)
Fix the ThemeMap and VerticalFullPageMapOrchestrator to support
`displayAllResults` in the noResults config of the map config. This
makes it so the toggle appears and the map updates to fit the
coordinates of the results.

J=SLAP-1232
TEST=manual

Test that when you have (below) in both the VerticalResults and
MapConfig noResults field, you will see the results display on a No
Results query (e.g. "askjdhakjsd"), the mobile view toggle will show and
the coordinates will fit on the map.
```
{
        "displayAllResults": true
}
```