Fix Hitchhiker universalsectiontemplate not being used for all verticals
This commit fixes a bug where the Hitchhiker universal section template
was only being used for verticals that had pages made for them. For example,
if the backend returned a universal section for vertical "fruits", but I did
not have a page created with verticalKey "fruits", the SDK universal section
template would be used instead of the Hitchhiker one.

This was fixed by overriding the "results/verticalresults" template.

J=SLAP-1130
TEST=manual

test that I can specify a specific universal section template
see that the hitchhiker template is always used even for verticals
by making only a universal page and seeing the hitchhiker template for
all verticls