Update RTF to support new elements (#730)
Update RTF to support strikethroughs, superscripts, subscripts, code, code blocks, blockquotes, and headings

The styling for these elements are the defaults for Chrome, with a few tweaks. The headings use `--yxt-font-size` values rather than the defaults, and the spacing is decreased so that it is more consistent with the KG spacing. The blockquote has a shorter indent because I think it works better on our cards. Finally, I made the code font smaller to make it more consistent with the default font.

J=SLAP-1273
TEST=manual

Use the Rose test account and the Answers rich text formatter on a card with all of the RTF that the KG supports