tweak spellcheck margins (#734)
This commit tweaks the spellcheck component's margin. It also moves it on top of
the results header on the full-page-map.

For universal standard, it was necessary to give spellcheck a negative margin.
This is because we want the spacing above the results to be 2.35 (in terms of yxt-base-spacing)
when there is no spellcheck, but only 1.5 when there is spellcheck (a decrease of 0.85).
It is not possible to use some kind of sibling selector, because the existence of spellcheck
is determined not by whether an Answers-spellcheck exists, but whether a yxt-Spellcheck
exists inside of an Answers-spellcheck. Sadly, css doesn't support a "parent" selector.
In this case, we would need a "has a sibling of Answers-spellcheck that is a parent of
a yxt-Spellcheck", which isn't possible.

J=SLAP-1130
TEST=manual,auto

test that on vertical-grid, vertical-standard, full-page-map, and universal-standard, the
spellcheck margin matches what Rose decided each should be (with and without a spellcheck
suggestion). Rose said she does not care about the old vertical-map page anymore.