consolidate webpack-config for locator-bundle and other bundles
This commit merges the locator-bundle's bespoke webpack config
into the config for the rest of the bundles. A following commit
will change the return value of the webpack-config to a single object
instead of an array. This refactor makes it easier to
modify the config based on whether the config is in development
or production mode.

J=SLAP-1132
TEST=manual

test that I can build the test-site, and the full-page-map and
regular page loads and can run searches