consolidate webpack-config for locator-bundle and other bundles (#738)
This commit merges the locator-bundle's bespoke webpack config
into the config for the rest of the bundles. A following commit
will change the return value of the webpack-config to a single object
instead of an array. This refactor makes it easier to
modify the config based on whether the config is in development
or production mode. This commit also renames bundle.js and bundle.css
to HitchhikerJS.js and HitchhikerCSS.css.

J=SLAP-1132
TEST=manual

test that I can build the test-site, and the full-page-map and
regular page loads and can run searches