add additional webpack optimizations for dev mode
This commit adds additional webpack optimizations when the
IS_DEVELOPMENT_PREVIEW env var is set to 'true'.
It also refactors the dev/prod mode logic into a decorator
class.

J=SLAP-1132
TEST=manual

test building the page with unset IS_DEVELOPMENT_PREVIEW,
IS_DEVELOPMENT_PREVIEW='false', and IS_DEVELOPMENT_PREVIEW='true',
and that the correct build mode is run for each

see that the build is much faster in development mode, and
that on ie11 I get a "Promise is not defined" error

see that ie11 still loads properly when in production mode,
and that assets are inlined