Fix locator bundle race condition (#739)
Fix a race condition where the vertical-full-page-map cards would break if the search loads before the locator bundle

J=SLAP-1276
TEST=manual

Use Fiddler to delay the loading of the locator bundle and observe the cards rendering properly. The cards aren't clickable while the bundle is loading, but as soon as it does, the cards become clickable. Also test that when the bundle loads before the search, the listeners register properly.