VerticalFullPageMap IE11 compatibility improvements (#742)
Update the VerticalFullPageMap to be fully IE11 compliant

We were using some syntax which wasn't supported by IE11. Update the locator to account for that. Also remove the flex on the mobile view toggle button because that was not being formatted properly on IE11. Removing the flex didn't change anything that I could see, and Percy didn't pick up any changes either.

Fixes for IE11 compatibility:
- For MediaQueryLists, use addListener rather than addEventListener
- Add a `removeElement()` helper function with IE11 support (essentially a ponyfill)
- Use setAttribute and removeAttribute for updating the detail card styling

J=none
TEST=manual

Test on IE11 and chrome and make sure that erros are no longer produced in the console on IE11, and test various functionality including pin clustering, opening and closing details cards, and toggling between mobile views.