product-prominentimage: fix ctas to bottom
This commit updates the product-prominentimage card
to always fix CTAs to the bottom of the card for the
vertical-grid layout. Previously, when cards on the same
row had different amounts of content above the CTA, for
example using different sized images, or shorter vs
longer card title text that may wrap, the CTAs would not
be aligned to the bottom of the card.

This was reported as a regression for 1.19, however I made a
site using theme 1.18, and this was still in issue there.
It's possible the Hitchhikers had some sort of internal
patch to get the desired behavior, and changing from 1.18 to 1.19
broke their patch.

J=SLAP-1192
TEST=manual

test that the CTAs are fixed to the bottom of the screen