use direct answer's timeZoneUtcOffset in openStatus
This commit updates the allfields standard direct
answer cards to use the timeZoneUtcOffset in the
direct answer. This timezone offset is the UTC offset
for the location that was returned as the direct answer.
Without this piece of information, the open status can be
formatted incorrectly if the location is in a different
timezone than the user.

Also fixed a bug with the global.Date mock we use for
percy snapshots and acceptance tests. The original
mock would return a pre-instantiated instance of
Date for usages where we try to get the current date.
The issue is that, we use the setHours and setMinutes
methods of Date in the HoursStringsLocalizer class,
which is used by the open status formatter. This would cause
that pre-instantiated instance of Date to be mutated.

The mocked date also now specifies GMT-0500 (EST) which
is what was already being assumed by my local machine,
on the off chance daylight savings time makes things weird.
It was already weird that the browser was assuming EST and
not EDT, when running new Date() in the console on regular
websites would display an EDT string.

J=SLAP-1270
TEST=manual,auto

added percy snapshot for direct answer using the open status formatter,
snapshot has 1 hr buffer for daylight savings time weirdness

checked that, on a universal page with without this change, the direct
answer open status, and the open status of a card displaying the same entity
could be out of sync, but with this change they would be in sync