add document-standard card
This card uses a copy-paste of the styling for the
standard card. Originally, I thought about adding
.HitchhikerDocumentStandard to standard.scss, instead
of making a new document-standard.scss, however I was
worried about the (albeit very unlikely) chance that someone
has already overrided standard.scss.

J=SLAP-1152
TEST=manual

check that the card works on ie11 and chrome, and looks like the mock