fix test-site's mock date
This commit fixes an issue with the mocked date we use
for our formatters in the test-site. Currently, we create
an instance of Date, and decorate the original Date class
to return this instance if new Date() is called with no params
(which should return the current date). However, the issue is that,
if new Date() is called, and then out of its setters is called,
like setHours() or setMinutes(), which we call in our HoursStringsLocalizer,
this fixed instance will be mutated. This causes inconsistent date time
behavior when using the test-site for local testing.

This is fixed by returning a new instance of Date every time the Date()
constructor is called.

J=SLAP-1313
TEST=manual

open dev tools and see that the output of new Date() is the expected date,
and that without this change it is not