dont remove .git folder in postupgrade script (#726)
Jambo now takes care of this, so we don't need to do it
in the postupgrade script. fs-extra's removeSync does not
error out if you try to remove a folder that does not exist,
so this change isn't strictly necessary. Thanks to this,
we also don't have to worry about Jambo backwards compatibility
with older theme versions.

J=SLAP-1135
TEST=manual

test upgrading to this branch with updated jambo, see .git is still removed