Remove the `node-sass` dependency from `static/package.json`.
The `node-sass` dependency in `static/package.json` is unused. When we first
created the Theme's toolchain, we copied much of the code from Consulting. The
Consulting code proved to be a lot more than we needed, so we did our best to
remove the unnecessary parts. This is likely a relic of the Consulting code
that we haven't removed yet.

TEST=manual

Ensured that I can still build my local test site and the styling looks as
expected. The `bundle.css` is unchanged.