Allow any SDK branch to be specified in `global_config.sdkVersion`. (#791)
This PR allows any SDK branch, not just official releases, to be specified in `global_config.sdkVersion`. To support this, a new HBS helper was added to the Theme: `sdkAssetUrl`. This helper takes in the branch (or release version), locale, and asset name, and provides the correct URL for it. With the addition of this helper, we were able to remove the `sdk-url` partial. The logic for computing the URL is now sufficiently complicated that we should do it in JS instead of HBS. 

J=SLAP-924
TEST=auto, manual

Added unit tests for the new helper. For the following scenarios, I tested with both an English and French locale, and verified the correct asset URLs were provided:
- Released SDK version.
- Release branch of the SDK.
- Feature branch of the SDK.
- Dev branch of the SDK.