Add runtime config support (#776)
Support runtime config and update the JWT support to use it

This PR allows the parent to set arbitrary config for iframed Answers experiences with the `setRuntimeConfig` function. Data attributes on the '#answers-container' div are also loaded into the RuntimeConfig when the iframe is created. Inside the experience, the runtime config is available at the `RuntimeConfig` object on the window.

The experience can be setup to allow be manually initialized rather than initialized automatically . To do this, the hitchhiker must set `"manuallyInitialize": true` inside global_config. With that option enabled, the experience must be started with `initAnswersExperienceFrame(config)` for iframe experiences or with `initAnswersExperience(config)` for non-iframe experiences.

The manual initialization also allows the experience to be started with a JWT token. To use JWT, 'initializeManually' and 'useJWT' should be set in global_config. The 'useJWT' option simply ensures that any api key that may have been injected does not appear anywhere in the build output. The experience can then be started with `initAnswersExperienceFrame({ token: 'jwt-token-here' });` or with `initAnswersExperience({ token: 'jwt-token-here' });` for non-iframe experiences.

J=SLAP-1315
TEST=manual

Test JWT support with iframe and non-iframe experiences. Test supplying custom config and seeing it available within the iframe. Note that the config is set even before the cards render with initial searches set. Test navigating pages within the iframe and seeing the config maintained. Set data attributes and see them get set within the iframe.