fix google streetview man being invisible
Mr. Pegman was being covered up due to our css
reset setting his max-height to 100%, which happened
to be 0 due to how the google maps js works.

This commit adds a css override on top of that.

T=409037
TEST=manual, auto

see mr. pegman again
add a percy snapshot for him (if it works?)