fix google streetview man being invisible (#801)
Mr. Pegman was being covered up due to our css
reset setting his max-height to 100%, which happened
to be 0 due to how the google maps js works.

This commit adds a css override on top of that.
I wanted to remove the actual line of the css reset
causing this behavior, however it's likely that removing
that line would cause ripples throughout the html-css continuum.

T=409037
TEST=manual, auto

see mr. pegman again
add a percy snapshot for him