fix PinImages and ClusterPinImages not changing pin size
One of the requirements of the vertical-full-page-map is
that users can jambo override the PinImages and ClusterPinImages
files in order to change the pin size. Rose says that this
was tested at some point, and since none of this code has been
changed since the initial 8000+ line full page map commit,
it's likely this was dropped somewhere during prototyping.

J=TECHOPS-603
TEST=manual

see that I can resize both the regular pin and the cluster pin,
on both chrome and ie11, for both google and mapbox