Use the esbuild-loader for production minification (#825)
Use the esbuild-loader in order to speed up our production build javascript and css minification

This changes speeds up the production webpack build by approximately 30 to 40%. The trade-off with esbuild is that the resulting minimized file is not quite as small. For index.html of the test site, the page size is about 2% larger.

We can remove cssnano and let esbuild handle css minification because the esbuild css minimizer leads to 8% faster webpack builds than CSS nano, and the resulting pages are 0.5% larger.

esbuild-loader does not support sourcemaps for css minification, however we currently don't provide css sourcemaps so this is not a regression

J=SLAP-1374
TEST=manual

Use the time program to measure the performance difference. Smoke test the site on chrome and on IE11.