change webpack devtool to 'eval' from 'source-map'
See here for a comparison of webpack devtool options
https://webpack.js.org/configuration/devtool/

'source-map' is listed as one of the slowest options
for build, and is recommended for production builds
that want high quality source maps.
Changing this to 'eval' shaves off 0.5s from the build time,
while still providing a source map in dev mode.

J=SLAP-1373
TEST=manual

see ~0.5s shaved off the webpack build time
see that I get source maps in dev mode, and can
add a console log to a formatter and clicking on
the console.log's line number will take me to the
correct source line in dev tools