Wrap 'Opens At' text in a call to `_localizer.getTranslation`. (#834)
In the `open-status/message-factory` we forgot to wrap one of the static strings in a `_localizer.getTranslation` call. This caused the 'Opens At' string to always appear in English.

J=SLAP-1380
TEST=manual, auto

This fix was already verified manually by a HH. I am just adding it to the Theme. I did add a unit test suite for the `message-factory`. The test suite and test I added would have caught this issue and so will ensure there aren't future regressions on this front.