add analytics to the video card (vimeo + youtube)
This commit adds onPlay analytics to the video card.
The vimeo and youtube iframe video apis are loaded
using code splitting, to not add to the bundle size
for non video card users, and also not block rendering.

J=SLAP-1381
TEST=manual

test that the video analytics are fired on play for
both vimeo and youtube

test clicking play multiple times

test in dev build mode and prod build mode, for ie11 and chrome

tried to add acceptance tests, but the video iframes do not play
well with testcafe. tried even a dummy page with just a single
hardcoded video iframe and I cannot click it in testcafe (either
programmatically or manually through Live mode

check that the vimeo and youtube apis are not loaded multiple
times even with multiple searches/cards)