Add the `product-prominentvideo` card to the Theme. (#847)
This PR adds the English-only variant of the `multilang-product-prominentvideo` card to the Theme. All calls to `translateJS` were replaced with static text. This PR also fixes the toggling of `Show More`/`Show Less` in both cards by adding `super.onMount` to the cards' custom `onMount`. 

TEST=manual

Verified I could use this card on a test vertical.