Add RTF link click analytics for Direct Answer cards.
This PR ensures that all Direct Answer cards in the Theme support RTF link
click analytics. The code was taken largely from the SDK. The main difference
is the source of some of the event options.

J=SLAP-1405
TEST=manual

Built a local test site with the new base `card_component`. Saw the correct
`AnalyticsEvent`s getting fired when I clicked on a RTF DA's links. Verified
that clicking plain text in the RTF value did not cause an event to be
fired.